feat(auth):#WB-608, style of "new device alert" email
