+++ auth/src/main/resources/i18n/fr.json
@@ -67,12 +67,12 @@
     "email.reset.pw.infos": "Vous avez demandé la réinitialisation du mot de passe.",
     "email.reset.id.infos": "Vous avez demandé un rappel de votre identifiant de connexion.",
     "email.signature" : "L'équipe Open Digital Education",
-    "email.new.device.subject": "Nouvelle connexion détectée",
-    "email.new.device.infos": "Une connexion a été réalisée sur votre compte {{#i18n}}project.name{{/i18n}} avec un nouvel appareil.",
-    "email.new.device.date": "<b>Date et heure de la connexion</b>: {{#datetime}}{{date}}{{/datetime}}",
-    "email.new.device.device": "<b>Appareil et navigateur</b>: {{device}}",
+    "email.new.device.subject": "Connexion sur un nouvel appareil",
+    "email.new.device.infos": "Afin de protéger votre compte, nous vous informons qu'une connexion à votre plateforme éducative a été effectuée sur un nouvel appareil.",
+    "email.new.device.date": "<b>Date et heure</b>: {{#datetime}}{{date}}{{/datetime}}",
+    "email.new.device.device": "<b>Appareil</b>: {{device}}",
     "email.new.device.ip": "<b>Adresse IP</b>: {{ip}}",
-    "email.new.device.warning": "Si vous n'êtes pas à l'origine de cette demande, nous vous conseillons de réinitialiser votre mot de passe.",
+    "email.new.device.warning": "S'il s'agit bien de vous, aucune action de votre part n'est requise. Dans le cas contraire, nous vous conseillons de <a href=\"{{host}}/auth/forgot#/password\" target=\"_blank\">modifier votre mot de passe.</a>",
     "phone.hi": "Bonjour, ",
     "phone.hi.withname": "Bonjour {{displayName}}, ",
     "phone.how.to.reset": "Vous pouvez vous connecter à l'ENT en saisissant votre identifiant et votre nouveau code d'activation",
